requirements-test.txt
flake8

pytest
pytest-env
pytest-django
pytest-pythonpath

coverage
django-webtest

pretend
